backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
python-multipart==0.0.9

# Database
supabase==2.5.0

# ML Core
torch>=2.1.0
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
hdbscan>=0.8.33
scikit-learn>=1.4.0
numpy>=2.0.0
pandas>=2.2.0
scipy>=1.12.0

# NLP / Zero-shot
transformers>=4.40.0
accelerate>=0.29.0

# Data validation
pydantic>=2.7.0
pydantic-settings>=2.3.0

# Auth / Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Utilities
httpx>=0.27.0
python-dotenv>=1.0.1
aiofiles>=23.2.1
loguru>=0.7.2
openpyxl>=3.1.2

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
